chore(PL-432): add missing technology relations on team seeds
